django/requirements.txt
Babel==2.16.0
boto3==1.35.70
caluma==10.6.0
caluma-alexandria==6.4.4
celery==5.4.0
cryptography==43.0.3
dataclasses-json==0.6.7
deepmerge==2.0
defusedxml==0.7.1
# check camac/admin.py MultilingualAdminMixin.get_search_result before updating
Django==4.2.16
django-admin-sortable2==2.2.4
django-clamd==1.0.0
django-environ==0.11.2
django-excel==0.0.10
django-filter==24.3
django-generic-api-permissions==0.4.7
django-hurricane==1.6.0
django-ipware==7.0.1
django-json-widget==2.0.1
# might be able to switch back to main fork https://github.com/Koed00/django-q/issues/735
django-q2==1.7.4
django-reversion==5.1.0
django-storages==1.14.4
# check camac/filter.py MultilingualSearchFilter.filter_queryset before updating
djangorestframework==3.15.2
djangorestframework-jsonapi==7.1.0
djangorestframework-xml==2.0.0
docxtpl==0.19.0
drf-extra-fields==3.7.0
drf-yasg==1.21.8
ebau-gwr==0.7.1
# We need to ensure we have this bugfix
# https://github.com/graphql-python/graphene/issues/1580
graphene==3.4.3
hdbcli==2.23.24
Jinja2==3.1.4
jwcrypto==1.5.6
lxml==5.3.0
manabi==1.5.0
mozilla-django-oidc==4.0.1
# openpyxl>3.0.10 is not compatible with pyexcel-xlsx 0.6.0
# https://github.com/pyexcel/pyexcel-xlsx/issues/52
openpyxl==3.0.10
Pillow==11.0.0
psycopg==3.2.3
pyexcel-io==0.6.7
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
PyJWT==2.10.0
pymemcache==4.0.0
pypdf==5.1.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-keycloak==4.7.3
python-dotenv==1.0.1
PyXB-X==1.2.6.2
qrcode==8.0
rapidfuzz==3.10.1
redis==5.2.0
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.6.0
# TODO: When updating sorl-thumbnail to a version that includes
# https://github.com/jazzband/sorl-thumbnail/pull/778, you can remove the
# monkeypatch in camac/document/apps.py
sorl-thumbnail==12.11.0
tabulate==0.9.0
tqdm==4.67.1
urllib3==2.2.3
uWSGI==2.0.28
whitenoise==6.8.2
WsgiDAV==4.3.3
xmlschema==3.4.3
